ebi-ait/dcp-ingest-central#517: combined the event handlers into a single file
